pyproject.toml
[tool.poetry]
name = "webpeditor"
version = "0.1.0"
description = ""
authors = ["Temirkhan Amanzhanov <tima1199@gmail.com>"]
license = "MIT License"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
dill = "^0.3.9"
lazy-object-proxy = "^1.10.0"
tomlkit = "^0.13.2"
platformdirs = "^4.3.6"
asgiref = "^3.8.1"
django-appconf = "^1.0.6"
dnspython = "^2.7.0"
mccabe = "^0.7.0"
pycodestyle = "^2.12.1"
pytz = "^2024.2"
toml = "^0.10.2"
tomli = "^2.0.2"
typing-extensions = "^4.12.2"
wrapt = "^1.16.0"
django-cors-headers = "^4.6.0"
django-environ = "^0.11.2"
django-extensions = "^3.2.3"
requests = "^2.32.3"
gunicorn = "^23.0.0"
whitenoise = "^6.8.2"
cloudinary = "^1.41.0"
pillow = "^11.0.0"
django-jsonfield = "^1.4.1"
deptry = "^0.21.0"
cryptography = "^43.0.3"
certifi = "^2024.8.30"
channels = {extras = ["daphne"], version = "^4.1.0"}
twisted = {extras = ["http2", "tls"], version = "^24.10.0"}
django = "^5.1.3"
sqlparse = "^0.5.1"
basedpyright = "^1.21.0"
django-compressor = "^4.5.1"
loguru = "^0.7.2"
returns = "^0.23.0"
django-injector = "^0.3.1"
djangorestframework = "^3.15.2"
djangorestframework-stubs = "^3.15.1"
django-crispy-forms = "^2.3"
crispy-tailwind = "^1.0.3"
drf-spectacular = {extras = ["sidecar"], version = "^0.27.2"}
djhtml = "^3.0.7"
django-jazzmin = "^3.0.1"
qrcode = "^8.0"
django-otp = "^1.5.4"
docutils = "^0.21.2"
django-cloudinary-storage = "^0.3.0"
django-types = "^0.19.1"

[tool.poetry.group.dev.dependencies]
ruff-lsp = "^0.0.57"
pip-audit = "^2.7.3"
pre-commit = "^4.0.1"
poetry-core = "^1.9.1"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 150
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "F401"]
ignore = ["E501", "B008"]
unfixable = ["B"]

[tool.basedpyright]
include = [
    "webpeditor",
    "webpeditor_app",
]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "**/.*",
    "htmlcov"
]

reportMissingImports = true
reportMissingTypeStubs = true
reportIncompatibleVariableOverride = false

pythonVersion = "3.12"
pythonPlatform = "Linux"

typeCheckingMode = "standard"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
